requirements.txt
# LINE Bot SDK
line-bot-sdk==3.9.0

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Google AI
google-generativeai==0.3.2

# Notion SDK
notion-client==2.2.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Image Processing
Pillow==10.1.0
requests==2.31.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[flask]==1.40.0

# Security
cryptography==41.0.8
authlib==1.3.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.8.0

# Production
psutil==5.9.6